[EI-236] [web] refactor: Refactor useIndicatorList hook and remove deprecated code
