+++ web/app/mocks/db/custom-forecast-indicator-action.mock.ts
@@ -93,7 +93,7 @@ export const mockCustomForecastIndicatorAction: MockCustomForecastIndicatorActio
       name: targetIndicator.name,
       ticker: targetIndicator.ticker,
       market: 'KOSPI',
-      type: 'k-stock',
+      type: 'customForecastIndicator',
       forecastType: 'multi',
       targetIndicatorValues: targetIndicatorValue.values,
       customForecastIndicatorValues: customForecastIndicatorValue.values,
