+++ web/app/store/querys/numerical-guidance/indicator.query.ts
@@ -2,8 +2,7 @@ import useSWR from 'swr';
 import { API_PATH } from '../api-path';
 import { fetchIndicatorsValue } from '../fetcher';
 import { Interval } from '../../stores/numerical-guidance/indicator-board.store';
-
-export type IndicatorType = 'customForecastIndicator' | 'k-stock';
+import { IndicatorType } from '../../stores/numerical-guidance/indicator-list.store';
 
 export type IndicatorsValueResponse = {
   indicatorsValue: IndicatorValueResponse[];
@@ -22,9 +21,19 @@ export type IndicatorValueItemResponse = {
   value: number | string;
 };
 
+// deprecate
 export const useFetchLiveIndicatorsValue = (indicatorIds: string[] | undefined, interval: Interval = 'day') => {
   // fix: type에 따라 url 변경 필요
   const key = indicatorIds ? [`${API_PATH.liveIndicatorValue}/k-stock`, interval, ...indicatorIds] : null;
 
   return useSWR<IndicatorsValueResponse, any, string[] | null>(key, fetchIndicatorsValue);
 };
+
+export type LiveIndicatorRequestParams = {
+  indicatorType: IndicatorType;
+  startDate: string;
+  interval: Interval;
+  ids: string[];
+};
+
+export const useFetchLiveIndicatorsValueByType = () => {};
